+++ games/highrise/src/main.ts
@@ -3,6 +3,7 @@ import { BootScene } from './scenes/BootScene'
 import { MenuScene } from './scenes/MenuScene'
 import { GameScene } from './scenes/GameScene'
 import { GameOverScene } from './scenes/GameOverScene'
+import { PauseScene } from './scenes/PauseScene'
 
 // Dynamic game canvas sizing.
 //
@@ -51,7 +52,7 @@ const config: Phaser.Types.Core.GameConfig = {
   input: {
     gamepad: true,
   },
-  scene: [BootScene, MenuScene, GameScene, GameOverScene],
+  scene: [BootScene, MenuScene, GameScene, GameOverScene, PauseScene],
 }
 
 new Phaser.Game(config)